refactor: Demo components location
