+++ _pages/dev/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { NextPage } from 'next';
 import { Stack, Typography } from '@mui/material';
-import DemoAppAlert from 'src/components/forPages/shared/DemoAppAlerts';
+import DemoAppAlert from 'app/dev/components/DemoAppAlerts';
 import DemoAppButton from 'src/components/forPages/shared/DemoAppButton';
 import DemoAppIcon from 'src/components/forPages/shared/DemoAppIcon';
 import DemoAppIconButton from 'src/components/forPages/shared/DemoAppIconButton';
