--- src/ReadMe.md
@@ -1,5 +1,0 @@
-# Why we need ./src folder at all?
-
-NextJS creates [routing URL for any source file](https://nextjs.org/docs/routing/introduction) in `./pages` folder :(
-
-So we need to host our sources, components and utils outside the "pages" folder, `./src` folder exactly for that
